requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
decouple==0.0.7
Django==5.0.2
django-cors-headers==4.3.1
django-decouple==2.1
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sentry-sdk==1.40.6
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
